pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "web-automation-agent"
version = "0.1.0"
description = "Deterministic web automation agent."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.52.0,<2.0.0",
    "pydantic>=2.11.0,<3.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "PyYAML>=6.0.2,<7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
